users app refactored. Implemented JWT utilities module under common package and separated JWT token generation and verification tasks
