+++ server/apps/users/views.py
@@ -1,15 +1,13 @@
-import jwt
-
-from django.conf import settings
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import ensure_csrf_cookie, csrf_protect
-from rest_framework.exceptions import AuthenticationFailed, NotAuthenticated
+from rest_framework.exceptions import AuthenticationFailed
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
-from .generate_jwt_tokens import generate_jwt_access_token, generate_jwt_refresh_token
+from ..common.jwt_utils import generate_jwt_access_token, generate_jwt_refresh_token
 from .models import User
 from .serializers import UserSerializer
+from ..common.jwt_utils import verify_jwt_token, verify_jwt_refresh_token
 
 
 class RegisterView(APIView):
@@ -65,12 +63,7 @@ class RefreshTokenView(APIView):
     def post(self, request):
         refresh_token = request.COOKIES.get('refresh_token')
 
-        if not refresh_token:
-            raise NotAuthenticated('Unauthenticated!')
-        try:
-            payload = jwt.decode(refresh_token, settings.REFRESH_TOKEN_SECRET, algorithms=['HS256'])
-        except jwt.ExpiredSignatureError:
-            raise NotAuthenticated('Authentication failed! Signature Expired!')
+        payload = verify_jwt_refresh_token(refresh_token)
 
         user = User.objects.filter(id=payload['id']).first()
 
@@ -93,12 +86,7 @@ class UserView(APIView):
     def get(request):
         token = request.COOKIES.get('jwt')
 
-        if not token:
-            raise NotAuthenticated('Unauthenticated!')
-        try:
-            payload = jwt.decode(token, settings.SECRET_KEY, algorithms=['HS256'])
-        except jwt.ExpiredSignatureError:
-            raise NotAuthenticated('Authentication failed! Signature Expired!')
+        payload = verify_jwt_token(token)
 
         user = User.objects.filter(id=payload['id']).first()
         serializer = UserSerializer(user)
